% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{bsol_montecarlo_WL3}
\alias{bsol_montecarlo_WL3}
\title{Monte Carlo simulation of waiting list}
\usage{
bsol_montecarlo_WL3(
  .data,
  seed,
  start_date_name = "start_date",
  end_date_name = "end_date",
  adds_name = "added",
  removes_name = "removed",
  starting_wl = 0L,
  run_id = 1L,
  force_boundary_alignment = TRUE,
  burn_in_days = 90L
)
}
\arguments{
\item{.data}{Data.frame.  Control data containing summarised data periods start and end dates,
'demand/referrals' ('adds'), 'capacity/removals' ('removes').}

\item{seed}{Numeric.   Specify a seed for reproducible outputs}

\item{start_date_name}{Character.  Name of the column containing start date.}

\item{end_date_name}{Character.  Name of the column containing end date.}

\item{adds_name}{Character.  Name of the column containing number of referrals/additions/demand.}

\item{removes_name}{Character.  Name of the column containing number of removals/capacity.}

\item{starting_wl}{Numeric.  The current waiting list size at the start of the period.}

\item{run_id}{When using in Monte Carlo, this would be 'i' for each replication. Default is 1 to run once.}

\item{force_boundary_alignment}{If TRUE, set the last day of each period’s size to the actual carry-over.}

\item{burn_in_days}{Number of days burn-in for initial backlog dispersion. Default is 90.}
}
\value{
A data.frame of dates (days) with waiting list size at each date.
}
\description{
A function, designed to be called from a loop, map or replicate, that allows repeated waiting list simulations,
passing the run_id and returning the daily waiting list in each run.  Process is stochastic, and designed to be different, using
random numbers drawn from distribution of your referrals and removals. This can then be averaged over for output.
}
\examples{
 # Monte Carlo control data
 control_data <-
  data.frame(
  start_date = as.Date(c("2024-04-01", "2024-07-01"), "\%Y-\%m-\%d"),
  end_date = as.Date(c("2024-06-30", "2024-09-30"), "\%Y-\%m-\%d"),
  Referrals = c(85, 112),
  Removals = c(90, 95)
  )

  sim_wls <-
     bsol_montecarlo_WL3(
         .data = control_data,
         start_date_name = "start_date",
         end_date_name = "end_date",
         adds_name = "Referrals",
         removes_name = "Removals",
         starting_wl = 450
     )

}
