% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_mean_wait_functions.R
\name{exceeding_prob}
\alias{exceeding_prob}
\title{Proportion exceeding target, given target and mean wait}
\usage{
exceeding_prob(mean_wait, threshold)
}
\arguments{
\item{mean_wait}{Mean waiting time (Wbar)}

\item{threshold}{numeric (>0). The waiting time threshold T (e.g., weeks).}
}
\value{
Numeric.
}
\description{
Given mean wait and threshold, return proportion exceeding target p = P(W > T)
using p = exp(- T / Wbar).
}
\examples{
exceeding_prob(12.94375, threshold = 52)
}
