% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_mean_wait_functions.R
\name{factor_for_exceeding}
\alias{factor_for_exceeding}
\title{Calculate factor argument, based on proportion exceeding target}
\usage{
factor_for_exceeding(p_exceed)
}
\arguments{
\item{p_exceed}{numeric in (0,1). Desired proportion exceeding T.}
}
\value{
Numeric. Factor (k)
}
\description{
Numeric factor used in average wait calculation - to get a quarter of
the target use factor=4 and one sixth of the target use factor = 6 etc. Defaults to 4.
see NHSRwaiting list
}
\details{
The "factor" k = T / mean = ln(1/p)
}
\examples{
factor_for_exceeding(0.018)
factor_for_exceeding(0.002)
factor_for_exceeding(0.05)
}
