% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_mean_wait_functions.R
\name{factor_for_meeting}
\alias{factor_for_meeting}
\title{Calculate the "factor", by proportion meeting target}
\usage{
factor_for_meeting(p_meeting)
}
\arguments{
\item{p_meeting}{numeric in (0,1). Desired proportion meeting T.}
}
\value{
Numeric. The 'factor' value (k)
}
\description{
Numeric factor used in average wait calculation - to get a quarter of
the target use factor=4 and one sixth of the target use factor = 6 etc. Defaults to 4.
see NHSRwaiting list
}
\details{
The "factor" k = T / mean = ln(1/p)
}
\examples{
factor_for_meeting(0.982)
factor_for_meeting(0.998)
factor_for_meeting(0.95)
}
