% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_mean_wait_functions.R
\name{mean_fraction_of_target}
\alias{mean_fraction_of_target}
\title{Mean wait as a proporiton of target wait.}
\usage{
mean_fraction_of_target(p_exceed)
}
\arguments{
\item{p_exceed}{numeric in (0,1). Desired proportion exceeding T.}
}
\value{
Numeric.  Mean as a fraction of the target (1/k)
}
\description{
Calculates the mean waiting time, based on exponential distribution, as a
fraction of target wait (T) = 1 / k
}
\examples{
mean_fraction_of_target(0.018) # 0.2489182
# So, applying to 52 weeks, mean wait for 0.018 exceeding (98.2 meeting)
# = 52 * 0.2489182
}
