% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_mean_wait_functions.R
\name{threshold_for_exceeding}
\alias{threshold_for_exceeding}
\title{Calculate the waiting times target from given mean wait and proportion exceeding}
\usage{
threshold_for_exceeding(mean_wait, p_exceed)
}
\arguments{
\item{mean_wait}{Mean waiting time (Wbar)}

\item{p_exceed}{numeric in (0,1). Desired proportion exceeding T.}
}
\value{
numeric. The target waiting time (T)
}
\description{
Given mean wait and desired proportion exceeding, return the threshold T
using T = - Wbar * ln(p_exceed).
}
\examples{
# P4 example: target should be T = 52 weeks
threshold_for_exceeding(12.943748, 0.018)
threshold_for_exceeding(8.367382, 0.002)
threshold_for_exceeding(17.358026, 0.05)


}
