% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_mean_wait_functions.R
\name{threshold_for_meeting}
\alias{threshold_for_meeting}
\title{Time threshold taken}
\usage{
threshold_for_meeting(mean_wait, p_meeting)
}
\arguments{
\item{mean_wait}{Mean waiting time (Wbar)}

\item{p_meeting}{numeric in (0,1). Desired proportion meeting T.}
}
\value{
numeric. The target waiting time (T)
}
\description{
Given mean wait and desired proportion meting target, return the threshold T
using T = - Wbar * ln(1-p_meeting).
}
\examples{
# P4 example: target should be T = 52 weeks
threshold_for_meeting(12.943748, 0.982)
threshold_for_meeting(8.367382, 0.998)
threshold_for_meeting(17.358026, 0.95)

}
