% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wl_join_cpp}
\alias{wl_join_cpp}
\title{Join two waiting lists}
\usage{
wl_join_cpp(wl_1, wl_2, referral_index = 0L)
}
\arguments{
\item{wl_1}{First waiting list}

\item{wl_2}{Second waiting list}

\item{referral_index}{Integer index of referral column (0-based in C++)}
}
\value{
Joined waiting list
}
\description{
Join two waiting lists
}
